Refactored MainApp into useMode
